+++ noodles-cram/examples/cram_write.rs
@@ -5,7 +5,7 @@
 //!
 //! Verify the output by piping to `samtools view --no-PG --with-header`.
 
-use std::io;
+use std::{io, num::NonZeroUsize};
 
 use noodles_core::Position;
 use noodles_cram as cram;
@@ -48,7 +48,9 @@ fn build_header(
 
     for record in reference_sequence_records {
         let name = record.name().parse()?;
-        let reference_sequence = Map::<ReferenceSequence>::new(record.sequence().len())?;
+        let length = NonZeroUsize::try_from(record.sequence().len())
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+        let reference_sequence = Map::<ReferenceSequence>::new(length);
         builder = builder.add_reference_sequence(name, reference_sequence);
     }